@@ -1,7 +1,10 @@
 <?php
 
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
-use App\Aluno;
+use Illuminate\Support\Facades\Hash;
+use App\Models\Aluno;
 
 class AlunoSeeder extends Seeder
 {
@@ -12,6 +15,7 @@ class AlunoSeeder extends Seeder
      */
     public function run()
     {
-	factory(Aluno::class, 5)->create();
+	//factory(Aluno::class, 5)->create();
+        Aluno::factory()->count(5)->create();
     }
 }
