@@ -1,7 +1,10 @@
 <?php
 
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
-use App\Mensagem;
+use Illuminate\Support\Facades\Hash;
+use App\Models\Mensagem;
 
 class MensagemSeeder extends Seeder
 {
@@ -12,6 +15,6 @@ class MensagemSeeder extends Seeder
      */
     public function run()
     {
-	factory(Mensagem::class, 5)->create();
+        Mensagem::factory()->count(5)->create();
     }
 }
