@@ -2,25 +2,42 @@
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
 
-use App\Model;
-use Faker\Generator as Faker;
+namespace Database\Factories;
 
-/*           $table->id();
-			$table->dateTime('data_envio');
-			$table->string('disciplina');
-			$table->string('nome_arq');
-            $table->timestamps();
+use App\Models\Administrador;
+use App\Models\Professor;
+use App\Models\Apostila;
+use Illuminate\Database\Eloquent\Factories\Factory;
+use Illuminate\Support\Str;
+
+/*
 			$table->integer('professor_id')->unsigned()->nullable();
 			$table->foreign('professor_id')->references('id')->on('professors');
 			$table->integer('administrador_id')->unsigned()->nullable();
 			$table->foreign('administrador_id')->references('id')->on('administradors');*/
 
-$factory->define(App\Apostila::class, function (Faker $faker) {
-    $nomearq = $faker->lexify('??????.');
-    $extensao = $faker->fileExtension;
-    return [
-        'data_envio' => ($faker->dateTimeBetween('-10 years', 'now'))->format('d/m/Y'),
-        'disciplina' => $faker->lexify('Disciplina ?????'),
-        'nome_arq' => $nomearq.$extensao
-    ];
-});
+class ApostilaFactory extends Factory{
+
+    protected $model = \App\Models\Apostila::class;
+    public function definition()
+    {
+        // TODO: Implement definition() method.
+        return [
+            'data_envio' => ($this->faker->dateTimeBetween('-10 years', 'now'))->format('d/m/Y'),
+            'disciplina' => $this->faker->lexify('Disciplina ?????'),
+            'nome_arq' => $this->faker->lexify('??????.').$this->faker->fileExtension
+        ];
+    }
+
+    public function profapo(){
+        return $this->state([
+            'professor_id' => $this->faker->randomElement(Professor::pluck('id')->toArray()),
+        ]);
+    }
+
+    public function admapo(){
+        return $this->state([
+            'administrador_id' => $this->faker->randomElement(Administrador::pluck('id')->toArray()),
+        ]);
+    }
+}
