@@ -2,32 +2,43 @@
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
 
-use App\Model;
-use Faker\Generator as Faker;
+namespace Database\Factories;
 
-$factory->define(App\Professor::class, function (Faker $faker) {
-    return [
-        'data_nascimento' => ($faker->dateTimeBetween('-100 years', '-16 years'))->format('d/m/Y'),
-        'email' => $faker->unique()->safeEmail,
-        'nome' => $faker->name(),
-        'telefone' => $faker->phoneNumber(),
-        'ehVoluntario' => $faker->randomElement(['Sim', 'Não']),
-        'grauInstrucao' => $faker->randomElement(['Estudante de Graduação', 'Graduação', 'Pós Graduação', 'Mestrado', 'Doutorado']),
-        'dispoManha' => $faker->boolean(),
-        'dispoTarde' => $faker->boolean(),
-        'dispoNoite' => $faker->boolean(),
-        'atuaBiologia' => $faker->boolean(),
-        'atuaEspanhol' => $faker->boolean(),
-        'atuaFilosofia' => $faker->boolean(),
-        'atuaFisica' => $faker->boolean(),
-        'atuaGeografia' => $faker->boolean(),
-        'atuaHistoria' => $faker->boolean(),
-        'atuaIngles' => $faker->boolean(),
-        'atuaLiteratura' => $faker->boolean(),
-        'atuaMatematica' => $faker->boolean(),
-        'atuaPortugues' => $faker->boolean(),
-        'atuaQuimica' => $faker->boolean(),
-        'atuaRedacao' => $faker->boolean(),
-        'atuaSociologia' => $faker->boolean()
-    ];
-});
+use App\Models\Professor;
+use Illuminate\Database\Eloquent\Factories\Factory;
+use Illuminate\Support\Str;
+
+
+class ProfessorFactory extends Factory{
+
+    protected $model = \App\Models\Professor::class;
+
+    public function definition()
+    {
+        // TODO: Implement definition() method.
+        return [
+            'data_nascimento' => ($this->faker->dateTimeBetween('-100 years', '-16 years'))->format('d/m/Y'),
+            'email' => $this->faker->unique()->safeEmail,
+            'nome' => $this->faker->name(),
+            'telefone' => $this->faker->phoneNumber(),
+            'ehVoluntario' => $this->faker->randomElement(['Sim', 'Não']),
+            'grauInstrucao' => $this->faker->randomElement(['Estudante de Graduação', 'Graduação', 'Pós Graduação', 'Mestrado', 'Doutorado']),
+            'dispoManha' => $this->faker->boolean(),
+            'dispoTarde' => $this->faker->boolean(),
+            'dispoNoite' => $this->faker->boolean(),
+            'atuaBiologia' => $this->faker->boolean(),
+            'atuaEspanhol' => $this->faker->boolean(),
+            'atuaFilosofia' => $this->faker->boolean(),
+            'atuaFisica' => $this->faker->boolean(),
+            'atuaGeografia' => $this->faker->boolean(),
+            'atuaHistoria' => $this->faker->boolean(),
+            'atuaIngles' => $this->faker->boolean(),
+            'atuaLiteratura' => $this->faker->boolean(),
+            'atuaMatematica' => $this->faker->boolean(),
+            'atuaPortugues' => $this->faker->boolean(),
+            'atuaQuimica' => $this->faker->boolean(),
+            'atuaRedacao' => $this->faker->boolean(),
+            'atuaSociologia' => $this->faker->boolean()
+        ];
+    }
+}
