@@ -2,8 +2,10 @@
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
 
-use App\Turma;
-use Faker\Generator as Faker;
+namespace Database\Factories;
+
+use App\Models\Turma;
+use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Str;
 
 /*
@@ -16,12 +18,18 @@ use Illuminate\Support\Str;
 | model instances for testing / seeding your application's database.
 |
 */
+class TurmaFactory extends Factory{
 
-$factory->define(Turma::class, function (Faker $faker) {
-    return [
-        'dataInicio' => ($faker->dateTimeBetween('-10 years', 'now'))->format('d/m/Y'),
-        'dataFim' => ($faker->dateTimeBetween('-10 years', 'now'))->format('d/m/Y'),
-	'horario' => $faker->time($format = 'H:i:s', $max = '20:00:00'),
-        'nome' => $faker->lexify('Turma ?????')
-    ];
-});
+    protected $model = Turma::class;
+
+    public function definition()
+    {
+        // TODO: Implement definition() method.
+        return [
+            'dataInicio' => ($this->faker->dateTimeBetween('-10 years', 'now'))->format('d/m/Y'),
+            'dataFim' => ($this->faker->dateTimeBetween('-10 years', 'now'))->format('d/m/Y'),
+            'horario' => $this->faker->time($format = 'H:i:s', $max = '20:00:00'),
+            'nome' => $this->faker->lexify('Turma ?????')
+        ];
+    }
+}
