@@ -1,5 +1,7 @@
 <?php
 
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
 
 class DatabaseSeeder extends Seeder
@@ -12,10 +14,10 @@ class DatabaseSeeder extends Seeder
     public function run()
     {
         $this->call(AdministradorSeeder::class);
-        $this->call(AlunoSeeder::class);
         $this->call(MensagemSeeder::class);
-        $this->call(TurmaSeeder::class);
+        $this->call(ProfessorTurmaSeeder::class);
+        $this->call(AlunoSeeder::class);
         $this->call(ApostilaSeeder::class);
-        $this->call(ProfessorSeeder::class);
+
     }
 }
